pyproject.toml
[project]
name = "store-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosmtplib>=4.0.1",
    "chainlit>=2.5.5",
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "openai>=1.83.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "twilio>=9.6.2",
    "uvicorn>=0.34.3",
]
